fix file path of audiotools.
