+++ paddlespeech/audiotools/core/util.py
@@ -734,8 +734,8 @@ def default_collate(batch,
         if not all(len(elem) == elem_size for elem in it):
             raise RuntimeError(
                 "each element in list of batch should be of equal size")
-        transposed = list(zip(
-            *batch))  # It may be accessed twice, so we use a list.
+        transposed = list(
+            zip(*batch))  # It may be accessed twice, so we use a list.
 
         if isinstance(elem, tuple):
             return [
