+++ paddlespeech/audiotools/requirements.txt
@@ -2,5 +2,4 @@ ffmpeg-python
 ffmpy
 flatten_dict
 pyloudnorm
-pytest
 rich
