+++ tests/unit/audiotools/core/test_effects.py
@@ -8,7 +8,7 @@ import numpy as np
 import paddle
 import pytest
 
-from audio.audiotools import AudioSignal
+from paddlespeech.audiotools import AudioSignal
 
 
 def test_normalize():
