+++ tests/unit/audiotools/core/test_util.py
@@ -11,8 +11,8 @@ import numpy as np
 import paddle
 import pytest
 
-from audio.audiotools import util
-from audio.audiotools.core.audio_signal import AudioSignal
+from paddlespeech.audiotools import util
+from paddlespeech.audiotools.core.audio_signal import AudioSignal
 from paddlespeech.vector.training.seeding import seed_everything
